Refactor

+++ capi/capi-runtime/src/lang/mod.rs
@@ -67,7 +67,7 @@ fn set_alpha(i: usize, mem: &mut [u8]) {
     set_channel(i, offset, value, mem);
 }
 
-fn set_channel(base_addr: usize, offset: usize, value: u8, mem: &mut [u8]) {
+fn set_channel(base_addr: usize, offset: usize, value: usize, mem: &mut [u8]) {
     let mut data_stack = DataStack::new();
 
     data_stack.push(base_addr);
@@ -75,7 +75,7 @@ fn set_channel(base_addr: usize, offset: usize, value: u8, mem: &mut [u8]) {
 
     add(&mut data_stack);
 
-    data_stack.push(value as usize);
+    data_stack.push(value);
 
     store(&mut data_stack, mem);
 }
